Rename and move PullRequestsController
Rename PullRequestsController to HooksController, change #status to
#slack and #update to #github
